+++ app/src/cmd_module.c
@@ -7,6 +7,8 @@
 #include <zephyr/fs/littlefs.h>
 #include <stdio.h>
 #include <string.h>
+#include <zephyr/random/random.h>
+
 
 #include "fs_module.h"
 #include "cmd_module.h"
@@ -42,13 +44,12 @@ volatile bool cmd_module_ble_connected = false;
 extern struct k_msgq q_sample;
 extern int global_dev_status;
 extern struct fs_mount_t *mp_sd;
+struct healthypi_session_log_header_t healthypi_session_log_header_data;
 bool settings_log_data_enabled = false;
 int current_log_counter;
 
 int8_t data_pkt[272];
 uint8_t buf_log[1024]; // 56 bytes / session, 18 sessions / packet
-
-//struct healthypi_session_log_header_t healthypi_session_log_header;
 
 struct wiser_cmd_data_fifo_obj_t
 {
@@ -62,6 +63,98 @@ struct wiser_cmd_data_fifo_obj_t
 K_FIFO_DEFINE(cmd_data_fifo);
 
 struct wiser_cmd_data_fifo_obj_t cmd_data_obj;
+
+void write_header_to_new_file()
+{
+    struct fs_file_t file;
+    int rc;
+    char session_name[50] = "/SD:/";
+
+    char session_id_str[20];
+    sprintf(session_id_str, "%d", healthypi_session_log_header_data.session_id);
+    strcat(session_name, session_id_str);
+    strcat(session_name, ".csv");
+
+    char session_record_details[200] = "Session started at: ";
+    char session_record_time[2];
+
+    sprintf(session_record_time, "%d", healthypi_session_log_header_data.session_start_time.day);
+    strcat(session_record_details, session_record_time);
+
+    strcat(session_record_details, "/");
+
+    sprintf(session_record_time, "%d", healthypi_session_log_header_data.session_start_time.month);
+    strcat(session_record_details, session_record_time);
+
+    strcat(session_record_details, "/");
+
+    sprintf(session_record_time, "%d", healthypi_session_log_header_data.session_start_time.year);
+    strcat(session_record_details, session_record_time);
+
+    strcat(session_record_details, " ");
+
+    sprintf(session_record_time, "%d", healthypi_session_log_header_data.session_start_time.hour);
+    strcat(session_record_details, session_record_time);
+
+    strcat(session_record_details, ":");
+
+    sprintf(session_record_time, "%d", healthypi_session_log_header_data.session_start_time.minute);
+    strcat(session_record_details, session_record_time);
+
+    strcat(session_record_details, ":");
+
+    sprintf(session_record_time, "%d", healthypi_session_log_header_data.session_start_time.second);
+    strcat(session_record_details, session_record_time);
+    strcat(session_record_details, "\n");
+
+    char session_vital_header[100] = "ECG,PPG,RESP\n";
+
+    fs_file_t_init(&file);
+
+    rc = fs_open(&file, session_name, FS_O_CREATE | FS_O_RDWR);
+    if (rc < 0)
+    {
+        printk("FAIL: open %s: %d", session_name, rc);
+    }
+
+    rc = fs_write(&file, session_record_details, strlen(session_record_details));
+    rc = fs_write(&file, session_vital_header, strlen(session_vital_header));
+
+
+    rc = fs_close(&file);
+    rc = fs_sync(&file);
+
+    printf("Header written to file... %d\n", healthypi_session_log_header_data.session_id);
+}
+
+
+void set_current_session_log_id(uint8_t m_sec, uint8_t m_min, uint8_t m_hour, uint8_t m_day, uint8_t m_month, uint8_t m_year)
+{
+    //printk("m_sec %d m_min %d, m_hour %d m_day %d m_month %d m_year %d\n", m_sec, m_min, m_hour, m_day, m_month, m_year);
+    uint8_t second, minute, hour, day, month, year;
+    year = m_year;
+    month = m_month;
+    day = m_day;
+    hour = m_hour;
+    minute = m_min;
+    second = m_sec;
+
+    // update structure with new log start time
+    healthypi_session_log_header_data.session_start_time.year = year;
+    healthypi_session_log_header_data.session_start_time.month = month;
+    healthypi_session_log_header_data.session_start_time.day = day;
+    healthypi_session_log_header_data.session_start_time.hour = hour;
+    healthypi_session_log_header_data.session_start_time.minute = minute;
+    healthypi_session_log_header_data.session_start_time.second = second;
+
+    uint8_t rand[2];
+    sys_rand_get(rand, sizeof(rand));
+    healthypi_session_log_header_data.session_id = (rand[0] | (rand[1] << 8));
+    healthypi_session_log_header_data.session_size = 0;
+
+    printk("Header data for log file %d set\n", healthypi_session_log_header_data.session_id);
+}
+
 
 void update_session_size_in_header(uint16_t file_size, char *m_file_path)
 {
